+++ ThreeDSecure/src/main/java/com/braintreepayments/api/ThreeDSecureLauncher.java
@@ -1,5 +1,7 @@
 package com.braintreepayments.api;
 
+import android.os.TransactionTooLargeException;
+
 import androidx.activity.result.ActivityResultLauncher;
 import androidx.activity.result.ActivityResultRegistry;
 import androidx.annotation.NonNull;
@@ -14,20 +16,22 @@ public class ThreeDSecureLauncher {
             "com.braintreepayments.api.ThreeDSecure.RESULT";
     @VisibleForTesting
     ActivityResultLauncher<ThreeDSecureResult> activityLauncher;
+    private CardinalResultCallback callback;
 
-   public ThreeDSecureLauncher(@NonNull Fragment fragment,
-                               @NonNull CardinalResultCallback callback) {
-       this(fragment.getActivity().getActivityResultRegistry(), fragment.getViewLifecycleOwner(),
-               callback);
-   }
+    public ThreeDSecureLauncher(@NonNull Fragment fragment,
+                                @NonNull CardinalResultCallback callback) {
+        this(fragment.getActivity().getActivityResultRegistry(), fragment.getViewLifecycleOwner(),
+                callback);
+    }
 
-   public ThreeDSecureLauncher(@NonNull FragmentActivity activity,
-                               @NonNull CardinalResultCallback callback) {
-       this(activity.getActivityResultRegistry(), activity, callback);
-   }
+    public ThreeDSecureLauncher(@NonNull FragmentActivity activity,
+                                @NonNull CardinalResultCallback callback) {
+        this(activity.getActivityResultRegistry(), activity, callback);
+    }
 
     public ThreeDSecureLauncher(ActivityResultRegistry registry, LifecycleOwner lifecycleOwner,
                                 CardinalResultCallback callback) {
+        this.callback = callback;
         activityLauncher =
                 registry.register(THREE_D_SECURE_RESULT, lifecycleOwner,
                         new ThreeDSecureActivityResultContract(),
@@ -35,6 +39,19 @@ public class ThreeDSecureLauncher {
     }
 
     public void launch(ThreeDSecureResult threeDSecureResult) {
-        activityLauncher.launch(threeDSecureResult);
+        try {
+            activityLauncher.launch(threeDSecureResult);
+        } catch (RuntimeException runtimeException) {
+            Throwable exceptionCause = runtimeException.getCause();
+            if (exceptionCause instanceof TransactionTooLargeException) {
+                String errorMessage = "The 3D Secure response returned is too large to continue. "
+                        + "Please contact Braintree Support for assistance.";
+                BraintreeException threeDSecureResponseTooLargeError =
+                        new BraintreeException(errorMessage, runtimeException);
+                callback.onCardinalResult(new CardinalResult(threeDSecureResponseTooLargeError));
+            } else {
+                throw runtimeException;
+            }
+        }
     }
 }
