+++ ThreeDSecure/src/test/java/com/braintreepayments/api/ThreeDSecureClientUnitTest.java
@@ -6,11 +6,16 @@ import static junit.framework.Assert.assertTrue;
 import static junit.framework.TestCase.assertNull;
 import static junit.framework.TestCase.assertSame;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.fail;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.ArgumentMatchers.isNull;
+import static org.mockito.ArgumentMatchers.same;
+import static org.mockito.Mockito.calls;
 import static org.mockito.Mockito.doAnswer;
+import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
@@ -19,6 +24,7 @@ import static org.mockito.Mockito.when;
 
 import android.content.Intent;
 import android.net.Uri;
+import android.os.TransactionTooLargeException;
 
 import androidx.activity.result.ActivityResultRegistry;
 import androidx.appcompat.app.AppCompatActivity;
@@ -30,8 +36,11 @@ import androidx.lifecycle.LifecycleObserver;
 import com.cardinalcommerce.cardinalmobilesdk.models.CardinalActionCode;
 import com.cardinalcommerce.cardinalmobilesdk.models.ValidateResponse;
 
+import junit.framework.TestCase;
+
 import org.json.JSONException;
 import org.json.JSONObject;
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -44,11 +53,8 @@ import org.robolectric.RobolectricTestRunner;
 public class ThreeDSecureClientUnitTest {
 
     private FragmentActivity activity;
-    private ActivityResultRegistry resultRegistry;
-    private Lifecycle lifecycle;
     private ThreeDSecureAPI threeDSecureAPI;
 
-    private ThreeDSecureListener listener;
     private ThreeDSecureResultCallback threeDSecureResultCallback;
 
     private Configuration threeDSecureEnabledConfig;
@@ -60,7 +66,6 @@ public class ThreeDSecureClientUnitTest {
     public void beforeEach() throws JSONException {
         activity = mock(FragmentActivity.class);
         threeDSecureResultCallback = mock(ThreeDSecureResultCallback.class);
-        listener = mock(ThreeDSecureListener.class);
         threeDSecureAPI = mock(ThreeDSecureAPI.class);
 
         threeDSecureEnabledConfig = new TestConfigurationBuilder()
@@ -78,13 +83,156 @@ public class ThreeDSecureClientUnitTest {
 
         threeDSecureResult =
                 ThreeDSecureResult.fromJson(Fixtures.THREE_D_SECURE_V2_LOOKUP_RESPONSE);
-
-        resultRegistry = mock(ActivityResultRegistry.class);
-        when(activity.getActivityResultRegistry()).thenReturn(resultRegistry);
-
-        lifecycle = mock(Lifecycle.class);
-        when(activity.getLifecycle()).thenReturn(lifecycle);
     }
+
+    // region prepareLookup
+
+    @Test
+    public void prepareLookup_returnsValidLookupJSONString()
+            throws JSONException, BraintreeException {
+        CardinalClient cardinalClient = new MockCardinalClientBuilder()
+                .successReferenceId("fake-df")
+                .build();
+
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(threeDSecureEnabledConfig)
+                .build();
+
+        ThreeDSecureClient sut =
+                new ThreeDSecureClient(braintreeClient, cardinalClient,
+                        new ThreeDSecureAPI(braintreeClient));
+
+        ThreeDSecurePrepareLookupCallback callback = mock(ThreeDSecurePrepareLookupCallback.class);
+        sut.prepareLookup(activity, basicRequest, callback);
+
+        ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
+        verify(callback).onResult(same(basicRequest), captor.capture(), isNull());
+
+        String clientData = captor.getValue();
+        JSONObject lookup = new JSONObject(clientData);
+        Assert.assertEquals("encoded_auth_fingerprint", lookup.getString("authorizationFingerprint"));
+        Assert.assertEquals(lookup.getString("braintreeLibraryVersion"),
+                "Android-" + BuildConfig.VERSION_NAME);
+        Assert.assertEquals(lookup.getString("dfReferenceId"), "fake-df");
+        Assert.assertEquals(lookup.getString("nonce"), "a-nonce");
+
+        JSONObject clientMetaData = lookup.getJSONObject("clientMetadata");
+        Assert.assertEquals(clientMetaData.getString("requestedThreeDSecureVersion"), "2");
+        Assert.assertEquals(clientMetaData.getString("sdkVersion"), "Android/" + BuildConfig.VERSION_NAME);
+    }
+
+    @Test
+    public void prepareLookup_returnsValidLookupJSONString_whenCardinalSetupFails()
+            throws JSONException, BraintreeException {
+        CardinalClient cardinalClient = new MockCardinalClientBuilder()
+                .error(new Exception("cardinal error"))
+                .build();
+
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(threeDSecureEnabledConfig)
+                .build();
+
+        ThreeDSecureClient sut =
+                new ThreeDSecureClient(braintreeClient, cardinalClient,
+                        new ThreeDSecureAPI(braintreeClient));
+
+        ThreeDSecurePrepareLookupCallback callback = mock(ThreeDSecurePrepareLookupCallback.class);
+        sut.prepareLookup(activity, basicRequest, callback);
+
+        ArgumentCaptor<String> captor = ArgumentCaptor.forClass(String.class);
+        verify(callback).onResult(same(basicRequest), captor.capture(), (Exception) isNull());
+
+        String clientData = captor.getValue();
+        JSONObject lookup = new JSONObject(clientData);
+        Assert.assertEquals("encoded_auth_fingerprint", lookup.getString("authorizationFingerprint"));
+        Assert.assertEquals(lookup.getString("braintreeLibraryVersion"),
+                "Android-" + BuildConfig.VERSION_NAME);
+        Assert.assertEquals(lookup.getString("nonce"), "a-nonce");
+        assertFalse(lookup.has("dfReferenceId"));
+
+        JSONObject clientMetaData = lookup.getJSONObject("clientMetadata");
+        Assert.assertEquals(clientMetaData.getString("requestedThreeDSecureVersion"), "2");
+        Assert.assertEquals(clientMetaData.getString("sdkVersion"), "Android/" + BuildConfig.VERSION_NAME);
+    }
+
+    @Test
+    public void prepareLookup_initializesCardinal() throws BraintreeException {
+        CardinalClient cardinalClient = new MockCardinalClientBuilder()
+                .successReferenceId("fake-df")
+                .build();
+
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(threeDSecureEnabledConfig)
+                .build();
+
+        ThreeDSecureClient sut =
+                new ThreeDSecureClient(braintreeClient, cardinalClient,
+                        new ThreeDSecureAPI(braintreeClient));
+
+        ThreeDSecurePrepareLookupCallback callback = mock(ThreeDSecurePrepareLookupCallback.class);
+        sut.prepareLookup(activity, basicRequest, callback);
+
+        verify(cardinalClient).initialize(same(activity), same(threeDSecureEnabledConfig),
+                same(basicRequest), any(CardinalInitializeCallback.class));
+    }
+
+    @Test
+    public void prepareLookup_whenCardinalClientInitializeFails_forwardsError()
+            throws BraintreeException {
+        BraintreeException initializeRuntimeError = new BraintreeException("initialize error");
+        CardinalClient cardinalClient = new MockCardinalClientBuilder()
+                .initializeRuntimeError(initializeRuntimeError)
+                .build();
+
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(threeDSecureEnabledConfig)
+                .build();
+
+        ThreeDSecureClient sut =
+                new ThreeDSecureClient(braintreeClient, cardinalClient,
+                        new ThreeDSecureAPI(braintreeClient));
+
+        ThreeDSecurePrepareLookupCallback callback = mock(ThreeDSecurePrepareLookupCallback.class);
+        sut.prepareLookup(activity, basicRequest, callback);
+
+        verify(callback).onResult(null, null, initializeRuntimeError);
+    }
+
+    @Test
+    public void prepareLookup_withoutCardinalJWT_postsException() throws BraintreeException {
+        CardinalClient cardinalClient = new MockCardinalClientBuilder().build();
+        Configuration configuration = new TestConfigurationBuilder()
+                .threeDSecureEnabled(true)
+                .buildConfiguration();
+
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(configuration)
+                .build();
+
+        ThreeDSecureClient sut =
+                new ThreeDSecureClient(braintreeClient, cardinalClient,
+                        new ThreeDSecureAPI(braintreeClient));
+
+        ThreeDSecurePrepareLookupCallback callback = mock(ThreeDSecurePrepareLookupCallback.class);
+        sut.prepareLookup(activity, basicRequest, callback);
+
+        ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(Exception.class);
+        verify(callback).onResult(isNull(), isNull(),
+                captor.capture());
+
+        TestCase.assertTrue(captor.getValue() instanceof BraintreeException);
+        Assert.assertEquals(captor.getValue().getMessage(), "Merchant is not configured for 3DS 2.0. " +
+                "Please contact Braintree Support for assistance.");
+    }
+
+    // endregion
+
+    // region performVerification
 
     @Test
     public void performVerification_sendsAnalyticEvent() throws BraintreeException {
@@ -117,7 +265,6 @@ public class ThreeDSecureClientUnitTest {
 
         ThreeDSecureRequest request = new ThreeDSecureRequest();
         request.setNonce("a-nonce");
-        request.setVersionRequested(ThreeDSecureRequest.VERSION_2);
         request.setAmount("amount");
         request.setRequestedExemptionType(ThreeDSecureRequest.SECURE_CORPORATE);
 
@@ -156,7 +303,6 @@ public class ThreeDSecureClientUnitTest {
 
         ThreeDSecureRequest request = new ThreeDSecureRequest();
         request.setNonce("a-nonce");
-        request.setVersionRequested(ThreeDSecureRequest.VERSION_2);
         request.setAmount("amount");
 
         ThreeDSecurePostalAddress billingAddress = new ThreeDSecurePostalAddress();
@@ -199,7 +345,6 @@ public class ThreeDSecureClientUnitTest {
 
         ThreeDSecureRequest request = new ThreeDSecureRequest();
         request.setNonce("a-nonce");
-        request.setVersionRequested(ThreeDSecureRequest.VERSION_2);
         request.setAmount("amount");
 
         ThreeDSecurePostalAddress billingAddress = new ThreeDSecurePostalAddress();
@@ -213,7 +358,7 @@ public class ThreeDSecureClientUnitTest {
         sut.performVerification(activity, request, threeDSecureResultCallback);
 
         verify(threeDSecureResultCallback).onResult(any(ThreeDSecureResult.class),
-                (Exception) isNull());
+                isNull());
     }
 
     @Test
@@ -225,7 +370,6 @@ public class ThreeDSecureClientUnitTest {
         ThreeDSecureClient sut =
                 new ThreeDSecureClient(braintreeClient, cardinalClient,
                         threeDSecureAPI);
-        sut.setListener(listener);
 
         ThreeDSecureRequest request = new ThreeDSecureRequest();
         request.setAmount("5");
@@ -235,209 +379,419 @@ public class ThreeDSecureClientUnitTest {
         verify(threeDSecureResultCallback).onResult(isNull(), captor.capture());
         assertEquals("The ThreeDSecureRequest nonce and amount cannot be null",
                 captor.getValue().getMessage());
-    }
+    }   @Test
+    public void performVerification_initializesCardinal() throws BraintreeException {
+        CardinalClient cardinalClient = new MockCardinalClientBuilder()
+                .successReferenceId("df-reference-id")
+                .build();
 
-    @Test
-    public void performVerification_whenV1_throwsAnError() throws BraintreeException {
-        CardinalClient cardinalClient = new MockCardinalClientBuilder().build();
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
                 .configuration(threeDSecureEnabledConfig)
                 .build();
 
         ThreeDSecureClient sut =
                 new ThreeDSecureClient(braintreeClient, cardinalClient,
-                        threeDSecureAPI);
-        basicRequest.setVersionRequested(ThreeDSecureRequest.VERSION_1);
-        sut.performVerification(activity, basicRequest, threeDSecureResultCallback);
+                        new ThreeDSecureAPI(braintreeClient));
+        sut.performVerification(activity, basicRequest, mock(ThreeDSecureResultCallback.class));
 
-        ArgumentCaptor<BraintreeException> captor =
-                ArgumentCaptor.forClass(BraintreeException.class);
-        verify(threeDSecureResultCallback).onResult(isNull(), captor.capture());
-
-        BraintreeException error = captor.getValue();
-        String expectedMessage =
-                "3D Secure v1 is deprecated and no longer supported. See https://developer.paypal.com/braintree/docs/guides/3d-secure/client-side/android/v4 for more information.";
-        assertEquals(expectedMessage, error.getMessage());
+        verify(cardinalClient).initialize(same(activity), same(threeDSecureEnabledConfig),
+                same(basicRequest), any(CardinalInitializeCallback.class));
     }
 
     @Test
-    public void onActivityResult_whenResultNotOk_postsExceptionToCallback()
+    public void performVerification_whenCardinalClientInitializeFails_forwardsError()
             throws BraintreeException {
-        CardinalClient cardinalClient = new MockCardinalClientBuilder().build();
-        BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
-
-        ThreeDSecureClient sut =
-                new ThreeDSecureClient(braintreeClient, cardinalClient,
-                        threeDSecureAPI);
-
-        verifyNoMoreInteractions(braintreeClient);
-        sut.onActivityResult(AppCompatActivity.RESULT_CANCELED, new Intent(),
-                threeDSecureResultCallback);
-        verifyNoMoreInteractions(braintreeClient);
-
-        ArgumentCaptor<BraintreeException> captor =
-                ArgumentCaptor.forClass(BraintreeException.class);
-        verify(threeDSecureResultCallback).onResult(isNull(),
-                captor.capture());
-
-        Exception exception = captor.getValue();
-        assertTrue(exception instanceof UserCanceledException);
-        assertEquals("User canceled 3DS.", exception.getMessage());
-        assertFalse(((UserCanceledException) exception).isExplicitCancelation());
-    }
-
-    @Test
-    public void onBrowserSwitchResult_whenSuccessful_postsPayment() throws BraintreeException {
-        Uri uri = Uri.parse("http://demo-app.com")
-                .buildUpon()
-                .appendQueryParameter("auth_response",
-                        Fixtures.THREE_D_SECURE_AUTHENTICATION_RESPONSE)
+        BraintreeException initializeRuntimeError = new BraintreeException("initialize error");
+        CardinalClient cardinalClient = new MockCardinalClientBuilder()
+                .initializeRuntimeError(initializeRuntimeError)
                 .build();
 
-        BrowserSwitchResult browserSwitchResult =
-                new BrowserSwitchResult(BrowserSwitchStatus.SUCCESS, null, uri);
-
-        CardinalClient cardinalClient = new MockCardinalClientBuilder().build();
-        BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
-
-        ThreeDSecureClient sut =
-                new ThreeDSecureClient(braintreeClient, cardinalClient,
-                        threeDSecureAPI);
-        sut.setListener(listener);
-
-        sut.onBrowserSwitchResult(browserSwitchResult);
-
-        ArgumentCaptor<ThreeDSecureResult> captor =
-                ArgumentCaptor.forClass(ThreeDSecureResult.class);
-        verify(listener).onThreeDSecureSuccess(captor.capture());
-        verify(listener, never()).onThreeDSecureFailure(any(Exception.class));
-
-        ThreeDSecureResult result = captor.getValue();
-        CardNonce cardNonce = result.getTokenizedCard();
-        assertIsANonce(cardNonce.getString());
-        assertEquals("11", cardNonce.getLastTwo());
-        assertTrue(cardNonce.getThreeDSecureInfo().wasVerified());
-    }
-
-    @Test
-    public void onBrowserSwitchResult_whenSuccessful_sendAnalyticsEvents()
-            throws BraintreeException {
-        Uri uri = Uri.parse("http://demo-app.com")
-                .buildUpon()
-                .appendQueryParameter("auth_response",
-                        Fixtures.THREE_D_SECURE_AUTHENTICATION_RESPONSE)
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(threeDSecureEnabledConfig)
                 .build();
 
-        BrowserSwitchResult browserSwitchResult =
-                new BrowserSwitchResult(BrowserSwitchStatus.SUCCESS, null, uri);
-
-        CardinalClient cardinalClient = new MockCardinalClientBuilder().build();
-        BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
-
         ThreeDSecureClient sut =
                 new ThreeDSecureClient(braintreeClient, cardinalClient,
-                        threeDSecureAPI);
-        sut.setListener(listener);
-        sut.onBrowserSwitchResult(browserSwitchResult);
+                        new ThreeDSecureAPI(braintreeClient));
 
-        verify(braintreeClient).sendAnalyticsEvent(
-                "three-d-secure.verification-flow.liability-shifted.true");
-        verify(braintreeClient).sendAnalyticsEvent(
-                "three-d-secure.verification-flow.liability-shift-possible.true");
+        ThreeDSecureResultCallback callback = mock(ThreeDSecureResultCallback.class);
+        sut.performVerification(activity, basicRequest, callback);
+
+        verify(callback).onResult(null, initializeRuntimeError);
     }
 
     @Test
-    public void onBrowserSwitchResult_whenFailure_postsErrorWithResponse() throws Exception {
-        JSONObject json = new JSONObject();
-        json.put("success", false);
-
-        JSONObject errorJson = new JSONObject();
-        errorJson.put("message", "Failed to authenticate, please try a different form of payment.");
-        json.put("error", errorJson);
-
-        Uri uri = Uri.parse("https://.com?auth_response=" + json.toString());
-
-        BrowserSwitchResult browserSwitchResult =
-                new BrowserSwitchResult(BrowserSwitchStatus.SUCCESS, null, uri);
-        CardinalClient cardinalClient = new MockCardinalClientBuilder().build();
-        BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
-
-        ThreeDSecureClient sut =
-                new ThreeDSecureClient(braintreeClient, cardinalClient,
-                        threeDSecureAPI);
-        sut.setListener(listener);
-        sut.onBrowserSwitchResult(browserSwitchResult);
-
-        ArgumentCaptor<ErrorWithResponse> captor = ArgumentCaptor.forClass(ErrorWithResponse.class);
-        verify(listener).onThreeDSecureFailure(captor.capture());
-        verify(listener, never()).onThreeDSecureSuccess(any(ThreeDSecureResult.class));
-
-        ErrorWithResponse error = captor.getValue();
-        assertEquals(422, error.getStatusCode());
-        assertEquals("Failed to authenticate, please try a different form of payment.",
-                error.getMessage());
-    }
-
-    @Test
-    public void onBrowserSwitchResult_whenBrowserSwitchStatusCanceled_returnsExceptionToCallback()
+    public void performVerification_whenCardinalSetupCompleted_sendsAnalyticEvent()
             throws BraintreeException {
-        BrowserSwitchResult browserSwitchResult =
-                new BrowserSwitchResult(BrowserSwitchStatus.CANCELED, null, null);
+        CardinalClient cardinalClient = new MockCardinalClientBuilder()
+                .successReferenceId("df-reference-id")
+                .build();
 
-        CardinalClient cardinalClient = new MockCardinalClientBuilder().build();
-        BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(threeDSecureEnabledConfig)
+                .build();
 
         ThreeDSecureClient sut =
                 new ThreeDSecureClient(braintreeClient, cardinalClient,
-                        threeDSecureAPI);
-        sut.setListener(listener);
-        sut.onBrowserSwitchResult(browserSwitchResult);
+                        new ThreeDSecureAPI(braintreeClient));
+        sut.performVerification(activity, basicRequest, mock(ThreeDSecureResultCallback.class));
+
+        verify(braintreeClient).sendAnalyticsEvent(
+                "three-d-secure.cardinal-sdk.init.setup-completed");
+    }
+
+    @Test
+    public void performVerification_whenCardinalSetupFailed_sendsAnalyticEvent()
+            throws BraintreeException {
+        CardinalClient cardinalClient = new MockCardinalClientBuilder()
+                .error(new Exception("cardinal error"))
+                .build();
+
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(threeDSecureEnabledConfig)
+                .build();
+
+        ThreeDSecureClient sut =
+                new ThreeDSecureClient(braintreeClient, cardinalClient,
+                        new ThreeDSecureAPI(braintreeClient));
+        sut.performVerification(activity, basicRequest, mock(ThreeDSecureResultCallback.class));
+
+        verify(braintreeClient).sendAnalyticsEvent("three-d-secure.cardinal-sdk.init.setup-failed");
+    }
+
+    @Test
+    public void performVerification_withoutCardinalJWT_postsException() throws BraintreeException {
+        CardinalClient cardinalClient = new MockCardinalClientBuilder().build();
+
+        Configuration configuration = new TestConfigurationBuilder()
+                .threeDSecureEnabled(true)
+                .buildConfiguration();
+
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(configuration)
+                .build();
+
+        ThreeDSecureClient sut =
+                new ThreeDSecureClient(braintreeClient, cardinalClient,
+                        new ThreeDSecureAPI(braintreeClient));
+
+        ThreeDSecureResultCallback callback = mock(ThreeDSecureResultCallback.class);
+        sut.performVerification(activity, basicRequest, callback);
 
         ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(Exception.class);
+        verify(callback).onResult(isNull(), captor.capture());
+        TestCase.assertTrue(captor.getValue() instanceof BraintreeException);
+        Assert.assertEquals(captor.getValue().getMessage(), "Merchant is not configured for 3DS 2.0. " +
+                "Please contact Braintree Support for assistance.");
+    }
+
+    // endregion
+
+    // region continuePerformVerification
+    @Test
+    public void continuePerformVerification_whenAuthenticatingWithCardinal_sendsAnalyticsEvent()
+            throws JSONException, BraintreeException {
+        CardinalClient cardinalClient = new MockCardinalClientBuilder()
+                .successReferenceId("reference-id")
+                .build();
+
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(threeDSecureEnabledConfig)
+                .build();
+
+        ThreeDSecureClient sut =
+                new ThreeDSecureClient(braintreeClient, cardinalClient,
+                        new ThreeDSecureAPI(braintreeClient));
+
+        ThreeDSecureResult threeDSecureResult =
+                ThreeDSecureResult.fromJson(Fixtures.THREE_D_SECURE_V2_LOOKUP_RESPONSE);
+        sut.continuePerformVerification(threeDSecureResult, threeDSecureResultCallback);
+
+        verify(braintreeClient).sendAnalyticsEvent("three-d-secure.verification-flow.started");
+    }
+
+    @Test
+    public void continuePerformVerification_whenChallengeIsPresented_sendsAnalyticsEvent()
+            throws JSONException, BraintreeException {
+        CardinalClient cardinalClient = new MockCardinalClientBuilder()
+                .successReferenceId("reference-id")
+                .build();
+
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(threeDSecureEnabledConfig)
+                .returnUrlScheme("sample-return-url://")
+                .build();
+
+        ThreeDSecureClient sut =
+                new ThreeDSecureClient(braintreeClient, cardinalClient,
+                        new ThreeDSecureAPI(braintreeClient));
+
+        ThreeDSecureResult threeDSecureResult =
+                ThreeDSecureResult.fromJson(Fixtures.THREE_D_SECURE_LOOKUP_RESPONSE);
+        sut.continuePerformVerification(threeDSecureResult, threeDSecureResultCallback);
+
+        verify(braintreeClient).sendAnalyticsEvent(
+                "three-d-secure.verification-flow.challenge-presented.true");
+    }
+
+    @Test
+    public void continuePerformVerification_whenChallengeIsNotPresented_sendsAnalyticsEvent()
+            throws JSONException, BraintreeException {
+        CardinalClient cardinalClient = new MockCardinalClientBuilder()
+                .successReferenceId("reference-id")
+                .build();
+
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(threeDSecureEnabledConfig)
+                .build();
+
+        ThreeDSecureClient sut =
+                new ThreeDSecureClient(braintreeClient, cardinalClient,
+                        new ThreeDSecureAPI(braintreeClient));
+
+        ThreeDSecureResult threeDSecureResult =
+                ThreeDSecureResult.fromJson(Fixtures.THREE_D_SECURE_LOOKUP_RESPONSE_NO_ACS_URL);
+        sut.continuePerformVerification(threeDSecureResult, threeDSecureResultCallback);
+
+        verify(braintreeClient).sendAnalyticsEvent(
+                "three-d-secure.verification-flow.challenge-presented.false");
+    }
+
+    @Test
+    public void continuePerformVerification_whenChallengeIsNotPresented_returnsResult()
+            throws JSONException, BraintreeException {
+        CardinalClient cardinalClient = new MockCardinalClientBuilder()
+                .successReferenceId("reference-id")
+                .build();
+
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(threeDSecureEnabledConfig)
+                .build();
+
+        ThreeDSecureClient sut =
+                new ThreeDSecureClient(braintreeClient, cardinalClient,
+                        new ThreeDSecureAPI(braintreeClient));
+
+        ThreeDSecureResult threeDSecureResult =
+                ThreeDSecureResult.fromJson(Fixtures.THREE_D_SECURE_LOOKUP_RESPONSE_NO_ACS_URL);
+        sut.continuePerformVerification(threeDSecureResult, threeDSecureResultCallback);
+
+        verify(threeDSecureResultCallback).onResult(threeDSecureResult, null);
+    }
+
+    @Test
+    public void continuePerformVerification_when3DSVersionIsVersion2_sendsAnalyticsEvent()
+            throws JSONException, BraintreeException {
+        CardinalClient cardinalClient = new MockCardinalClientBuilder()
+                .successReferenceId("reference-id")
+                .build();
+
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(threeDSecureEnabledConfig)
+                .build();
+
+        ThreeDSecureClient sut =
+                new ThreeDSecureClient(braintreeClient, cardinalClient,
+                        new ThreeDSecureAPI(braintreeClient));
+
+        ThreeDSecureResult threeDSecureResult =
+                ThreeDSecureResult.fromJson(Fixtures.THREE_D_SECURE_V2_LOOKUP_RESPONSE);
+        sut.continuePerformVerification(threeDSecureResult, threeDSecureResultCallback);
+
+        verify(braintreeClient).sendAnalyticsEvent(
+                "three-d-secure.verification-flow.3ds-version.2.1.0");
+    }
+
+    @Test
+    public void continuePerformVerification_whenObserverIsNull_startsActivity()
+            throws JSONException, BraintreeException {
+        CardinalClient cardinalClient = new MockCardinalClientBuilder()
+                .successReferenceId("reference-id")
+                .build();
+
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(threeDSecureEnabledConfig)
+                .build();
+
+        ThreeDSecureClient sut =
+                new ThreeDSecureClient(braintreeClient, cardinalClient,
+                        new ThreeDSecureAPI(braintreeClient));
+        ThreeDSecureResult threeDSecureResult =
+                ThreeDSecureResult.fromJson(Fixtures.THREE_D_SECURE_V2_LOOKUP_RESPONSE);
+
+        sut.continuePerformVerification(threeDSecureResult,
+                mock(ThreeDSecureResultCallback.class));
+
+        verify(activity).startActivityForResult(any(Intent.class), any(Integer.class));
+    }
+
+    @Test
+    public void continuePerformVerification_whenObserverIsNullAndTransactionIsTooLarge_callsBackAnException()
+            throws JSONException, BraintreeException {
+        CardinalClient cardinalClient = new MockCardinalClientBuilder()
+                .successReferenceId("reference-id")
+                .build();
+
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(threeDSecureEnabledConfig)
+                .build();
+
+        ThreeDSecureClient sut =
+                new ThreeDSecureClient(braintreeClient, cardinalClient,
+                        new ThreeDSecureAPI(braintreeClient));
+        ThreeDSecureResult threeDSecureResult =
+                ThreeDSecureResult.fromJson(Fixtures.THREE_D_SECURE_V2_LOOKUP_RESPONSE);
+
+        TransactionTooLargeException transactionTooLargeException =
+                new TransactionTooLargeException();
+        RuntimeException runtimeException = new RuntimeException(
+                "runtime exception caused by transaction too large", transactionTooLargeException);
+
+        doThrow(runtimeException)
+                .when(activity).startActivityForResult(any(Intent.class), anyInt());
+
+        ThreeDSecureResultCallback callback = mock(ThreeDSecureResultCallback.class);
+        sut.continuePerformVerification(threeDSecureResult, callback);
+
+        ArgumentCaptor<BraintreeException> captor =
+                ArgumentCaptor.forClass(BraintreeException.class);
+        verify(callback).onResult(isNull(), captor.capture());
+
+        BraintreeException braintreeException = captor.getValue();
+        String expectedMessage = "The 3D Secure response returned is too large to continue. "
+                + "Please contact Braintree Support for assistance.";
+        Assert.assertEquals(expectedMessage, braintreeException.getMessage());
+    }
+
+    @Test
+    public void continuePerformVerification_whenObserverIsNullAndRuntimeExceptionThrown_rethrowsException()
+            throws JSONException, BraintreeException {
+        CardinalClient cardinalClient = new MockCardinalClientBuilder()
+                .successReferenceId("reference-id")
+                .build();
+
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(threeDSecureEnabledConfig)
+                .build();
+
+        ThreeDSecureClient sut =
+                new ThreeDSecureClient(braintreeClient, cardinalClient,
+                        new ThreeDSecureAPI(braintreeClient));
+        ThreeDSecureResult threeDSecureResult =
+                ThreeDSecureResult.fromJson(Fixtures.THREE_D_SECURE_V2_LOOKUP_RESPONSE);
+
+        RuntimeException runtimeException = new RuntimeException("random runtime exception");
+        doThrow(runtimeException)
+                .when(activity).startActivityForResult(any(Intent.class), anyInt());
+
+        ThreeDSecureResultCallback callback = mock(ThreeDSecureResultCallback.class);
+
+        try {
+            sut.continuePerformVerification(threeDSecureResult, callback);
+            fail("should not get here");
+        } catch (Exception e) {
+            Assert.assertSame(e, runtimeException);
+        }
+    }
+
+    @Test
+    public void continuePerformVerification_withObserverAndTransactionIsTooLarge_callsBackAnException()
+            throws JSONException, BraintreeException {
+        CardinalClient cardinalClient = new MockCardinalClientBuilder()
+                .successReferenceId("reference-id")
+                .build();
+
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(threeDSecureEnabledConfig)
+                .build();
+
+        ThreeDSecureClient sut =
+                new ThreeDSecureClient(braintreeClient, cardinalClient,
+                        new ThreeDSecureAPI(braintreeClient));
+
+        ThreeDSecureResult threeDSecureResult =
+                ThreeDSecureResult.fromJson(Fixtures.THREE_D_SECURE_V2_LOOKUP_RESPONSE);
+        TransactionTooLargeException transactionTooLargeException =
+                new TransactionTooLargeException();
+        RuntimeException runtimeException = new RuntimeException(
+                "runtime exception caused by transaction too large", transactionTooLargeException);
+        doThrow(runtimeException).when(threeDSecureLifecycleObserver).launch(threeDSecureResult);
+        sut.observer = threeDSecureLifecycleObserver;
+
+        sut.continuePerformVerification(basicRequest, threeDSecureResult);
+        ArgumentCaptor<BraintreeException> captor =
+                ArgumentCaptor.forClass(BraintreeException.class);
         verify(listener).onThreeDSecureFailure(captor.capture());
-        verify(listener, never()).onThreeDSecureSuccess(any(ThreeDSecureResult.class));
 
-        Exception exception = captor.getValue();
-        assertTrue(exception instanceof UserCanceledException);
-        assertEquals("User canceled 3DS.", exception.getMessage());
-        assertFalse(((UserCanceledException) exception).isExplicitCancelation());
+        BraintreeException braintreeException = captor.getValue();
+        String expectedMessage = "The 3D Secure response returned is too large to continue. "
+                + "Please contact Braintree Support for assistance.";
+        Assert.assertEquals(expectedMessage, braintreeException.getMessage());
     }
 
     @Test
-    public void onBrowserSwitchResult_whenListenerNull_setsPendingBrowserSwitchResult_andDoesNotDeliver()
+    public void continuePerformVerification_withObserverAndRuntimeExceptionThrown_rethrowsException()
+            throws JSONException, BraintreeException {
+        CardinalClient cardinalClient = new MockCardinalClientBuilder()
+                .successReferenceId("reference-id")
+                .build();
+
+        BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
+                .authorizationSuccess(Authorization.fromString(Fixtures.BASE64_CLIENT_TOKEN))
+                .configuration(threeDSecureEnabledConfig)
+                .build();
+
+        ThreeDSecureClient sut =
+                new ThreeDSecureClient(braintreeClient, cardinalClient,
+                        new ThreeDSecureAPI(braintreeClient));
+
+        ThreeDSecureResult threeDSecureResult =
+                ThreeDSecureResult.fromJson(Fixtures.THREE_D_SECURE_V2_LOOKUP_RESPONSE);
+        RuntimeException runtimeException = new RuntimeException("random runtime exception");
+        doThrow(runtimeException).when(threeDSecureLifecycleObserver).launch(threeDSecureResult);
+        sut.observer = threeDSecureLifecycleObserver;
+
+        try {
+            sut.continuePerformVerification(basicRequest, threeDSecureResult);
+            fail("should not get here");
+        } catch (Exception e) {
+            Assert.assertSame(e, runtimeException);
+        }
+    }
+
+    // endregion
+
+    // region onCardinalResult
+    @Test
+    public void onCardinalResult_whenErrorExists_forwardsErrorToCallback_andSendsAnalytics()
             throws BraintreeException {
         CardinalClient cardinalClient = new MockCardinalClientBuilder().build();
-
-        BrowserSwitchResult browserSwitchResult = mock(BrowserSwitchResult.class);
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
 
         ThreeDSecureClient sut =
                 new ThreeDSecureClient(braintreeClient, cardinalClient,
                         threeDSecureAPI);
-        sut.onBrowserSwitchResult(browserSwitchResult);
-
-        assertSame(browserSwitchResult, sut.pendingBrowserSwitchResult);
-        verify(listener, never()).onThreeDSecureFailure(any(Exception.class));
-        verify(listener, never()).onThreeDSecureSuccess(any(ThreeDSecureResult.class));
-    }
-
-    @Test
-    public void onCardinalResult_whenErrorExists_forwardsErrorToListener_andSendsAnalytics()
-            throws BraintreeException {
-        CardinalClient cardinalClient = new MockCardinalClientBuilder().build();
-        BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
-
-        ThreeDSecureClient sut =
-                new ThreeDSecureClient(braintreeClient, cardinalClient,
-                        threeDSecureAPI);
-        ThreeDSecureListener listener = mock(ThreeDSecureListener.class);
-        sut.setListener(listener);
 
         Exception threeDSecureError = new Exception("3DS error.");
         CardinalResult cardinalResult = new CardinalResult(threeDSecureError);
-        sut.onCardinalResult(cardinalResult, );
+        sut.onCardinalResult(cardinalResult, threeDSecureResultCallback);
 
-        verify(listener).onThreeDSecureFailure(threeDSecureError);
+        ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(Exception.class);
+        verify(threeDSecureResultCallback).onResult(isNull(), captor.capture());
     }
 
     @Test
@@ -451,12 +805,10 @@ public class ThreeDSecureClientUnitTest {
         ThreeDSecureClient sut =
                 new ThreeDSecureClient(braintreeClient, cardinalClient,
                         threeDSecureAPI);
-        ThreeDSecureListener listener = mock(ThreeDSecureListener.class);
-        sut.setListener(listener);
 
         CardinalResult cardinalResult =
                 new CardinalResult(threeDSecureResult, "jwt", validateResponse);
-        sut.onCardinalResult(cardinalResult, );
+        sut.onCardinalResult(cardinalResult, threeDSecureResultCallback);
 
         verify(braintreeClient).sendAnalyticsEvent(
                 "three-d-secure.verification-flow.cardinal-sdk.action-code.success");
@@ -475,15 +827,13 @@ public class ThreeDSecureClientUnitTest {
         ThreeDSecureClient sut =
                 new ThreeDSecureClient(braintreeClient, cardinalClient,
                         threeDSecureAPI);
-        ThreeDSecureListener listener = mock(ThreeDSecureListener.class);
-        sut.setListener(listener);
 
         CardinalResult cardinalResult =
                 new CardinalResult(threeDSecureResult, "jwt", validateResponse);
-        sut.onCardinalResult(cardinalResult, );
+        sut.onCardinalResult(cardinalResult, threeDSecureResultCallback);
 
         ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(Exception.class);
-        verify(listener).onThreeDSecureFailure(captor.capture());
+        verify(threeDSecureResultCallback).onResult(isNull(), captor.capture());
         assertTrue(captor.getValue() instanceof BraintreeException);
         assertEquals("Error", captor.getValue().getMessage());
 
@@ -502,15 +852,13 @@ public class ThreeDSecureClientUnitTest {
         ThreeDSecureClient sut =
                 new ThreeDSecureClient(braintreeClient, cardinalClient,
                         threeDSecureAPI);
-        ThreeDSecureListener listener = mock(ThreeDSecureListener.class);
-        sut.setListener(listener);
 
         CardinalResult cardinalResult =
                 new CardinalResult(threeDSecureResult, "jwt", validateResponse);
-        sut.onCardinalResult(cardinalResult, );
+        sut.onCardinalResult(cardinalResult, threeDSecureResultCallback);
 
         ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(Exception.class);
-        verify(listener).onThreeDSecureFailure(captor.capture());
+        verify(threeDSecureResultCallback).onResult(isNull(), captor.capture());
         Exception exception = captor.getValue();
         assertTrue(exception instanceof UserCanceledException);
         assertEquals("User canceled 3DS.", exception.getMessage());
@@ -528,28 +876,23 @@ public class ThreeDSecureClientUnitTest {
         ValidateResponse validateResponse = mock(ValidateResponse.class);
         when(validateResponse.getActionCode()).thenReturn(CardinalActionCode.SUCCESS);
 
-        doAnswer(new Answer<Void>() {
-            @Override
-            public Void answer(InvocationOnMock invocation) {
-                ThreeDSecureResultCallback callback =
-                        (ThreeDSecureResultCallback) invocation.getArguments()[2];
-                callback.onResult(threeDSecureResult, null);
-                return null;
-            }
+        doAnswer((Answer<Void>) invocation -> {
+            ThreeDSecureResultCallback callback =
+                    (ThreeDSecureResultCallback) invocation.getArguments()[2];
+            callback.onResult(threeDSecureResult, null);
+            return null;
         }).when(threeDSecureAPI).authenticateCardinalJWT(any(ThreeDSecureResult.class), anyString(),
                 any(ThreeDSecureResultCallback.class));
 
         ThreeDSecureClient sut =
                 new ThreeDSecureClient(braintreeClient, cardinalClient,
                         threeDSecureAPI);
-        ThreeDSecureListener listener = mock(ThreeDSecureListener.class);
-        sut.setListener(listener);
 
         CardinalResult cardinalResult =
                 new CardinalResult(threeDSecureResult, "jwt", validateResponse);
-        sut.onCardinalResult(cardinalResult, );
+        sut.onCardinalResult(cardinalResult, threeDSecureResultCallback);
 
-        verify(listener).onThreeDSecureSuccess(threeDSecureResult);
+        verify(threeDSecureResultCallback).onResult(threeDSecureResult, null);
         verify(braintreeClient).sendAnalyticsEvent(
                 "three-d-secure.verification-flow.upgrade-payment-method.succeeded");
         verify(braintreeClient).sendAnalyticsEvent(
@@ -571,28 +914,23 @@ public class ThreeDSecureClientUnitTest {
         final ThreeDSecureResult threeDSecureResult = mock(ThreeDSecureResult.class);
         when(threeDSecureResult.hasError()).thenReturn(true);
 
-        doAnswer(new Answer<Void>() {
-            @Override
-            public Void answer(InvocationOnMock invocation) {
-                ThreeDSecureResultCallback callback =
-                        (ThreeDSecureResultCallback) invocation.getArguments()[2];
-                callback.onResult(threeDSecureResult, null);
-                return null;
-            }
+        doAnswer((Answer<Void>) invocation -> {
+            ThreeDSecureResultCallback callback =
+                    (ThreeDSecureResultCallback) invocation.getArguments()[2];
+            callback.onResult(threeDSecureResult, null);
+            return null;
         }).when(threeDSecureAPI).authenticateCardinalJWT(any(ThreeDSecureResult.class), anyString(),
                 any(ThreeDSecureResultCallback.class));
 
         ThreeDSecureClient sut =
                 new ThreeDSecureClient(braintreeClient, cardinalClient,
                         threeDSecureAPI);
-        ThreeDSecureListener listener = mock(ThreeDSecureListener.class);
-        sut.setListener(listener);
 
         CardinalResult cardinalResult =
                 new CardinalResult(threeDSecureResult, "jwt", validateResponse);
-        sut.onCardinalResult(cardinalResult, );
+        sut.onCardinalResult(cardinalResult, threeDSecureResultCallback);
 
-        verify(listener).onThreeDSecureSuccess(threeDSecureResult);
+        verify(threeDSecureResultCallback).onResult(threeDSecureResult, null);
         verify(braintreeClient).sendAnalyticsEvent(
                 "three-d-secure.verification-flow.upgrade-payment-method.failure.returned-lookup-nonce");
         verify(braintreeClient).sendAnalyticsEvent("three-d-secure.verification-flow.completed");
@@ -609,207 +947,27 @@ public class ThreeDSecureClientUnitTest {
 
         final Exception exception = new Exception("error");
 
-        doAnswer(new Answer<Void>() {
-            @Override
-            public Void answer(InvocationOnMock invocation) {
-                ThreeDSecureResultCallback callback =
-                        (ThreeDSecureResultCallback) invocation.getArguments()[2];
-                callback.onResult(null, exception);
-                return null;
-            }
+        doAnswer((Answer<Void>) invocation -> {
+            ThreeDSecureResultCallback callback =
+                    (ThreeDSecureResultCallback) invocation.getArguments()[2];
+            callback.onResult(null, exception);
+            return null;
         }).when(threeDSecureAPI).authenticateCardinalJWT(any(ThreeDSecureResult.class), anyString(),
                 any(ThreeDSecureResultCallback.class));
 
         ThreeDSecureClient sut =
                 new ThreeDSecureClient(braintreeClient, cardinalClient,
                         threeDSecureAPI);
-        ThreeDSecureListener listener = mock(ThreeDSecureListener.class);
-        sut.setListener(listener);
 
         CardinalResult cardinalResult =
                 new CardinalResult(threeDSecureResult, "jwt", validateResponse);
-        sut.onCardinalResult(cardinalResult, );
+        sut.onCardinalResult(cardinalResult, threeDSecureResultCallback);
 
-        verify(listener).onThreeDSecureFailure(exception);
+        verify(threeDSecureResultCallback).onResult(null, exception);
         braintreeClient.sendAnalyticsEvent(
                 "three-d-secure.verification-flow.upgrade-payment-method.errored");
         verify(braintreeClient).sendAnalyticsEvent("three-d-secure.verification-flow.completed");
     }
 
-    @Test
-    public void constructor_setsLifecycleObserver() throws BraintreeException {
-        CardinalClient cardinalClient = new MockCardinalClientBuilder().build();
-        BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
-
-        ThreeDSecureClient sut =
-                new ThreeDSecureClient(braintreeClient, cardinalClient,
-                        threeDSecureAPI);
-
-        ArgumentCaptor<ThreeDSecureLifecycleObserver> captor =
-                ArgumentCaptor.forClass(ThreeDSecureLifecycleObserver.class);
-        verify(lifecycle).addObserver(captor.capture());
-
-        ThreeDSecureLifecycleObserver observer = captor.getValue();
-        assertSame(resultRegistry, observer.activityResultRegistry);
-        assertSame(sut, observer.threeDSecureClient);
-    }
-
-    @Test
-    public void constructor_withFragment_passesFragmentLifecycleAndActivityToObserver() {
-        BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
-
-        Fragment fragment = mock(Fragment.class);
-        when(fragment.requireActivity()).thenReturn(activity);
-        when(fragment.getLifecycle()).thenReturn(lifecycle);
-        ThreeDSecureClient sut = new ThreeDSecureClient(fragment, braintreeClient);
-
-        ArgumentCaptor<ThreeDSecureLifecycleObserver> captor =
-                ArgumentCaptor.forClass(ThreeDSecureLifecycleObserver.class);
-        verify(lifecycle).addObserver(captor.capture());
-
-        ThreeDSecureLifecycleObserver observer = captor.getValue();
-        assertSame(resultRegistry, observer.activityResultRegistry);
-        assertSame(sut, observer.threeDSecureClient);
-    }
-
-    @Test
-    public void constructor_withFragmentActivity_passesActivityLifecycleAndActivityToObserver() {
-        BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
-
-        when(activity.getLifecycle()).thenReturn(lifecycle);
-
-        ThreeDSecureClient sut = new ThreeDSecureClient(activity, braintreeClient);
-
-        ArgumentCaptor<ThreeDSecureLifecycleObserver> captor =
-                ArgumentCaptor.forClass(ThreeDSecureLifecycleObserver.class);
-        verify(lifecycle).addObserver(captor.capture());
-
-        ThreeDSecureLifecycleObserver observer = captor.getValue();
-        assertSame(resultRegistry, observer.activityResultRegistry);
-        assertSame(sut, observer.threeDSecureClient);
-    }
-
-    @Test
-    public void constructor_withoutFragmentOrActivity_doesNotSetObserver() {
-        BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
-
-        ThreeDSecureClient sut = new ThreeDSecureClient(braintreeClient);
-
-        verify(lifecycle, never()).addObserver(any(LifecycleObserver.class));
-    }
-
-    @Test
-    public void setListener_whenPendingBrowserSwitchResultExists_deliversResultToListener_andSetsPendingResultNull()
-            throws BraintreeException {
-        CardinalClient cardinalClient = new MockCardinalClientBuilder().build();
-
-        Uri uri = Uri.parse("http://demo-app.com")
-                .buildUpon()
-                .appendQueryParameter("auth_response",
-                        Fixtures.THREE_D_SECURE_AUTHENTICATION_RESPONSE)
-                .build();
-
-        BrowserSwitchResult browserSwitchResult =
-                new BrowserSwitchResult(BrowserSwitchStatus.SUCCESS, null, uri);
-        BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
-
-        ThreeDSecureClient sut =
-                new ThreeDSecureClient(braintreeClient, cardinalClient,
-                        threeDSecureAPI);
-        sut.pendingBrowserSwitchResult = browserSwitchResult;
-        sut.setListener(listener);
-
-        verify(listener).onThreeDSecureSuccess(any(ThreeDSecureResult.class));
-        verify(listener, never()).onThreeDSecureFailure(any(Exception.class));
-
-        assertNull(sut.pendingBrowserSwitchResult);
-    }
-
-    @Test
-    public void setListener_whenPendingBrowserSwitchResultDoesNotExist_doesNotInvokeListener()
-            throws BraintreeException {
-        CardinalClient cardinalClient = new MockCardinalClientBuilder().build();
-        BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
-
-        ThreeDSecureClient sut =
-                new ThreeDSecureClient(braintreeClient, cardinalClient,
-                        threeDSecureAPI);
-        sut.pendingBrowserSwitchResult = null;
-        sut.setListener(listener);
-
-        verify(listener, never()).onThreeDSecureSuccess(any(ThreeDSecureResult.class));
-        verify(listener, never()).onThreeDSecureFailure(any(Exception.class));
-
-        assertNull(sut.pendingBrowserSwitchResult);
-    }
-
-    @Test
-    public void getBrowserSwitchResult_forwardsInvocationToBraintreeClient()
-            throws BraintreeException {
-        CardinalClient cardinalClient = new MockCardinalClientBuilder().build();
-        BraintreeClient braintreeClient = mock(BraintreeClient.class);
-
-        BrowserSwitchResult browserSwitchResult = mock(BrowserSwitchResult.class);
-        when(braintreeClient.getBrowserSwitchResult(activity)).thenReturn(browserSwitchResult);
-
-        ThreeDSecureClient sut =
-                new ThreeDSecureClient(braintreeClient, cardinalClient,
-                        threeDSecureAPI);
-
-        BrowserSwitchResult result = sut.getBrowserSwitchResult(activity);
-        assertSame(browserSwitchResult, result);
-    }
-
-    @Test
-    public void deliverBrowserSwitchResult_forwardsInvocationToBraintreeClient()
-            throws BraintreeException {
-        CardinalClient cardinalClient = new MockCardinalClientBuilder().build();
-        BraintreeClient braintreeClient = mock(BraintreeClient.class);
-
-        BrowserSwitchResult browserSwitchResult = mock(BrowserSwitchResult.class);
-        when(braintreeClient.deliverBrowserSwitchResult(activity)).thenReturn(browserSwitchResult);
-
-        ThreeDSecureClient sut =
-                new ThreeDSecureClient(braintreeClient, cardinalClient,
-                        threeDSecureAPI);
-
-        BrowserSwitchResult result = sut.deliverBrowserSwitchResult(activity);
-        assertSame(browserSwitchResult, result);
-    }
-
-    @Test
-    public void getBrowserSwitchResultFromCache_forwardsInvocationToBraintreeClient()
-            throws BraintreeException {
-        CardinalClient cardinalClient = new MockCardinalClientBuilder().build();
-        BraintreeClient braintreeClient = mock(BraintreeClient.class);
-
-        BrowserSwitchResult browserSwitchResult = mock(BrowserSwitchResult.class);
-        when(braintreeClient.getBrowserSwitchResultFromNewTask(activity)).thenReturn(
-                browserSwitchResult);
-
-        ThreeDSecureClient sut =
-                new ThreeDSecureClient(braintreeClient, cardinalClient,
-                        threeDSecureAPI);
-
-        BrowserSwitchResult result = sut.getBrowserSwitchResultFromNewTask(activity);
-        assertSame(browserSwitchResult, result);
-    }
-
-    @Test
-    public void deliverBrowserSwitchResultFromNewTask_forwardsInvocationToBraintreeClient()
-            throws BraintreeException {
-        CardinalClient cardinalClient = new MockCardinalClientBuilder().build();
-        BraintreeClient braintreeClient = mock(BraintreeClient.class);
-
-        BrowserSwitchResult browserSwitchResult = mock(BrowserSwitchResult.class);
-        when(braintreeClient.deliverBrowserSwitchResultFromNewTask(activity)).thenReturn(
-                browserSwitchResult);
-
-        ThreeDSecureClient sut =
-                new ThreeDSecureClient(braintreeClient, cardinalClient,
-                        threeDSecureAPI);
-
-        BrowserSwitchResult result = sut.deliverBrowserSwitchResultFromNewTask(activity);
-        assertSame(browserSwitchResult, result);
-    }
+    // endregion
 }
